server: Split out the PaRoT Options
